13. 리뷰 쓰기(갤러리 카메라) 리뷰 불러오기 (완)

+++ app/src/main/java/hgh/project/baemin_clone/widget/adapter/viewholder/review/RestaurantReviewViewHolder.kt
@@ -37,7 +37,7 @@ class RestaurantReviewViewHolder(
             }
             reviewTitleText.text=model.title
             reviewText.text= model.description
-            ratingBar.rating =model.grade.toFloat()
+            ratingBar.rating =model.grade
         }
 
     }
